+++ internal/core/user/impl/kratos/config/identity.schema.json
