+++ internal/core/user/impl/uc/uc-adaptor/dao/uc_sync_record.go
