+++ internal/core/user/impl/uc/uc-adaptor/endpoints/info.go
