+++ internal/core/user/impl/uc/uc-adaptor/ucclient/ucclient.go
@@ -26,8 +26,8 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/internal/core/user/uc"
-	"github.com/erda-project/erda/internal/core/user/uc-adaptor/conf"
+	"github.com/erda-project/erda/internal/core/user/impl/uc"
+	"github.com/erda-project/erda/internal/core/user/impl/uc/uc-adaptor/conf"
 	"github.com/erda-project/erda/pkg/discover"
 	"github.com/erda-project/erda/pkg/http/httpclient"
 	"github.com/erda-project/erda/pkg/strutil"
