+++ internal/tools/openapi/legacy/api/apis/uc/uc_user_export_test.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/internal/core/user/uc"
+	"github.com/erda-project/erda/internal/core/user/impl/uc"
 )
 
 func TestGetLoginMethodMap(t *testing.T) {
